Fix edge case in isNewKeyForTarget
isNewKeyForTarget is used to notify watchers
of the target object that depend on e.g. the
Object.keys(target) value.

The previous implementation of this check used
`typeof target[key] === "undefined"`, which registers
a false positive when `key` exists in the object,
but has the literal value `undefined`.

Instead we can use `!(key in target)` to prevent
unwanted rerenders in this case.