Fix components rendered via another components children being re-created
In a case where a component was rendered within another component using
the `children` property of the parent component, the child component
would be re-created on every render. This was caused due to the props
being deeply wrapped in state.

To alleviate this problem `shallow` flag was introduced to the
`createState` function which only wraps the top-most layer into a proxy.
This might be useful for users as well, so it is exposed as a part of
the public API.