Remove node context objects
This change was achieved by first making the `parentElement_` of each
node instance mandatory. This does require the introduction of a new
`hasMounted` field, but allows us to deduce the namespace of HTML
elements without needing to pass context objects around.

Additionally, the `parentElement_` need not be a `HtmlElementInstance`
but can instead refer to the raw DOM `Element` instead. This removes
more indirection.